Rename adapter file.
