+++ smarts/core/utils/adapters/ego_centric_adapters.py
